Pequeño cambio en migracion de offers El número de vacantes puede ser un entero, en vez de un double.

+++ wishten/database/migrations/2023_08_15_154817_create_offers_table.php
@@ -19,10 +19,10 @@ return new class extends Migration
             $table->string('title');
             $table->double('salary');
             $table->string('description');
-            $table->double('vacants');
+            $table->integer('vacants');
             $table->string('document_path');
             $table->timestamps(); //crea automaticamnte 2 columnas, una de created_at y una de updated_at
-            
+
         });
     }
